out of merge_hell
